+++ x/btcstaking/types/query.go
@@ -70,7 +70,7 @@ func (ud *BTCUndelegation) ToResponse() (resp *BTCUndelegationResponse) {
 }
 
 // NewFinalityProviderResponse creates a new finality provider response based on the finality provider
-func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight uint64) *FinalityProviderResponse {
+func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight uint64, softDeleted bool) *FinalityProviderResponse {
 	return &FinalityProviderResponse{
 		Description:          f.Description,
 		Commission:           f.Commission,
@@ -83,5 +83,6 @@ func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight uint64) *Fi
 		Height:               bbnBlockHeight,
 		HighestVotedHeight:   f.HighestVotedHeight,
 		CommissionInfo:       f.CommissionInfo,
+		SoftDeleted:          softDeleted,
 	}
 }